First commit of the copy website
Basic layout and flex boxes made. Still needs quite a bit of
improvement but we're close. The quote section, cta, footer,
header and section1 all need work. In particular alligning
elements in the header and section1, centering in the section3
the quote, centering in section4 the cta, pushing the footer to
the bottom of the webpage, fixing combo1, and I'm sure there is
more, but it's a decent start.