Sumbit site
Not perfect there are a lot of things that i'm sure could have been done
better or differently, but when compared to the image of the site we were
supposed to copy it's quite close. If I were to have a better understanding
and ability I'd fix sections growig so they all were matched and do less
padding bottom properties, I'd fix the button in the cta-box. (it's sizing
is a bit off) and getting the site to fill to the edges. But I'm moving on.